refactor: move actor details DTOs to model package
