+++ data/src/main/java/com/london/data/datasource/remote/details/actordetails/model/ActorTvShowDetailsDto.kt
@@ -1,4 +1,4 @@
-package com.london.data.datasource.remote.details.actordetails
+package com.london.data.datasource.remote.details.actordetails.model
 
 import kotlinx.serialization.SerialName
