+++ data/src/main/java/com/london/data/datasource/remote/search/RemoteDataSource.kt
@@ -1,8 +1,8 @@
 package com.london.data.datasource.remote.search
 
-import com.london.data.datasource.remote.details.actordetails.ActorDetailsDto
-import com.london.data.datasource.remote.details.actordetails.ActorMovieDetailsDto
-import com.london.data.datasource.remote.details.actordetails.ActorTvShowDetailsDto
+import com.london.data.datasource.remote.details.actordetails.model.ActorDetailsDto
+import com.london.data.datasource.remote.details.actordetails.model.ActorMovieDetailsDto
+import com.london.data.datasource.remote.details.actordetails.model.ActorTvShowDetailsDto
 import com.london.data.datasource.remote.search.model.ApiSearch
 import com.london.data.datasource.remote.search.model.SearchActorRemote
 import com.london.data.datasource.remote.search.model.SearchMovieRemote
